NMS-12259: Unify tests
